Rename files with integers less than 10 to include leading 0 for proper lexical sorting.
