+++ src/App.js
@@ -11,15 +11,15 @@ import ReduxTestComponent from './test-components/ReduxTestComponent'
 // And change the App component to render the correct test component
 
 // import React Challenges:
-import * as React_1 from './challenges/react/curriculum/React_1'
-import * as React_2 from './challenges/react/curriculum/React_2'
-import * as React_3 from './challenges/react/curriculum/React_3'
-import * as React_4 from './challenges/react/curriculum/React_4'
-import * as React_5 from './challenges/react/curriculum/React_5'
-import * as React_6 from './challenges/react/curriculum/React_6'
-import * as React_7 from './challenges/react/curriculum/React_7'
-import * as React_8 from './challenges/react/curriculum/React_8'
-import * as React_9 from './challenges/react/curriculum/React_9'
+import * as React_01 from './challenges/react/curriculum/React_01'
+import * as React_02 from './challenges/react/curriculum/React_02'
+import * as React_03 from './challenges/react/curriculum/React_03'
+import * as React_04 from './challenges/react/curriculum/React_04'
+import * as React_05 from './challenges/react/curriculum/React_05'
+import * as React_06 from './challenges/react/curriculum/React_06'
+import * as React_07 from './challenges/react/curriculum/React_07'
+import * as React_08 from './challenges/react/curriculum/React_08'
+import * as React_09 from './challenges/react/curriculum/React_09'
 import * as React_10 from './challenges/react/curriculum/React_10'
 import * as React_11 from './challenges/react/curriculum/React_11'
