+++ src/challenges/react/curriculum/React_06.js
